requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1.3
django-extensions==3.2.3
django-simple-history==3.7.0
django-versatileimagefield==3.1
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
pillow==11.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-magic==0.4.27
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.21.0
sqlparse==0.5.1
uritemplate==4.1.1
